requirements.txt
python-dotenv
matplotlib
boto3
transformers
torch